adsf

+++ src/main/java/ignis/ignis/domain/feed/service/FeedService.java
@@ -120,6 +120,9 @@ public class FeedService {
     @Transactional
     public void deleteFeed(Long feedId) {
         Feed feed = feedRepository.findById(feedId).orElseThrow(()-> new RuntimeException("s"));
+        if (!userFacade.getCurrentUser().getUserName().equals(feed.getUser().getUserName())) {
+            throw new RuntimeException("asdf");
+        }
         feedRepository.delete(feed);
     }
 }
